// src/app/tinkoff-api/model/candleResolution.ts
/**
 * OpenAPI
 * tinkoff.ru/invest OpenAPI.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: n.v.melnikov@tinkoff.ru
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Интервал свечи и допустимый промежуток запроса: - 1min [1 minute, 1 day] - 2min [2 minutes, 1 day] - 3min [3 minutes, 1 day] - 5min [5 minutes, 1 day] - 10min [10 minutes, 1 day] - 15min [15 minutes, 1 day] - 30min [30 minutes, 1 day] - hour [1 hour, 7 days] - day [1 day, 1 year] - week [7 days, 2 years] - month [1 month, 10 years] 
 */
export type CandleResolution = '1min' | '2min' | '3min' | '5min' | '10min' | '15min' | '30min' | 'hour' | 'day' | 'week' | 'month';

export const CandleResolution = {
    _1min: '1min' as CandleResolution,
    _2min: '2min' as CandleResolution,
    _3min: '3min' as CandleResolution,
    _5min: '5min' as CandleResolution,
    _10min: '10min' as CandleResolution,
    _15min: '15min' as CandleResolution,
    _30min: '30min' as CandleResolution,
    Hour: 'hour' as CandleResolution,
    Day: 'day' as CandleResolution,
    Week: 'week' as CandleResolution,
    Month: 'month' as CandleResolution
};