// src/app/tinkoff-api/model/candle.ts
/**
 * OpenAPI
 * tinkoff.ru/invest OpenAPI.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: n.v.melnikov@tinkoff.ru
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CandleResolution } from './candleResolution';

export interface Candle { 
    figi: string;
    interval: CandleResolution;
    o: number;
    c: number;
    h: number;
    l: number;
    v: number;
    /**
     * ISO8601
     */
    time: Date;
}